Major re-design based on the Bootstrap framework.
This version of xs3p now:

 - Generates modern, UTF-8 encoded HTML5 output
 - Has a fresh look & feel based on the Bootstrap framework
 - Supports markdown on `<documentation>` elements

This re-design was major overhaul and some features may have gone missing or
may not yet be completely ported, but the majority of functionality (plus more)
is readily available. For example are links from the `source` attribute of the
`<documentation>` attribute now rendered in the documentation section.

The page layout may still have a few bugs for certain inputs. I am still trying
to iron out some very minor issues with the content menu, for example. But all
output is valid HTML5 and properly displayed in all major modern browsers.

You can find a rendering of the example input that I specifically added to show
the capabilities of this fork here:

https://bitfehler.net/xs3p/address_annotated.xsd.html